pyproject.toml
[project]
name = "mcp-integration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.86.0",
    "arxiv>=2.4.1",
    "mcp>=1.26.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.2",
]
